started making tests for routes, tests for routes/main.py
